# Coursework/Databases/Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Shop Backend API
  version: 1.0.0
  description: API локального backend-сервиса кофейни.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/ingredient:
    get:
      tags: [ingredient]
      summary: Получить список ингредиентов
      parameters:
        - in: query
          name: include_inactive
          schema:
            type: boolean
            default: true
          description: Включать неактивные ингредиенты
      responses:
        '200':
          description: Список ингредиентов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListIngredientsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [ingredient]
      summary: Создать ингредиент
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIngredientRequest'
      responses:
        '200':
          description: Ингредиент создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIngredientResponse'
        '400':
          oneOf:
            - $ref: '#/components/responses/ValidationError'
            - $ref: '#/components/responses/ShiftNotActive'
        '409':
          $ref: '#/components/responses/DuplicateName'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/ingredient/{ingredient_id}:
    get:
      tags: [ingredient]
      summary: Получить ингредиент
      parameters:
        - $ref: '#/components/parameters/IngredientID'
      responses:
        '200':
          description: Карточка ингредиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIngredientResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/ingredient/{ingredient_id}/activity:
    patch:
      tags: [ingredient]
      summary: Изменить активность ингредиента
      parameters:
        - $ref: '#/components/parameters/IngredientID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetIngredientActivityRequest'
      responses:
        '200':
          description: Активность ингредиента обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetIngredientActivityResponse'
        '400':
          $ref: '#/components/responses/IngredientUsedInRecipe'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/menu-item:
    get:
      tags: [menu-item]
      summary: Получить список пунктов меню
      parameters:
        - in: query
          name: include_inactive
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Список пунктов меню
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMenuItemsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [menu-item]
      summary: Создать пункт меню
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemRequest'
      responses:
        '200':
          description: Пункт меню создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMenuItemResponse'
        '400':
          oneOf:
            - $ref: '#/components/responses/ValidationError'
            - $ref: '#/components/responses/InactiveIngredient'
        '409':
          $ref: '#/components/responses/DuplicateName'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/menu-item/{menu_item_id}:
    get:
      tags: [menu-item]
      summary: Получить пункт меню
      parameters:
        - $ref: '#/components/parameters/MenuItemID'
      responses:
        '200':
          description: Карточка пункта меню
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMenuItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/menu-item/{menu_item_id}/price:
    patch:
      tags: [menu-item]
      summary: Обновить цену пункта меню
      parameters:
        - $ref: '#/components/parameters/MenuItemID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemPriceRequest'
      responses:
        '200':
          description: Цена обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateMenuItemPriceResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/order:
    get:
      tags: [order]
      summary: Получить список заказов
      parameters:
        - in: query
          name: employee_id
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrdersResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [order]
      summary: Создать заказ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          oneOf:
            - $ref: '#/components/responses/ValidationError'
            - $ref: '#/components/responses/ShiftNotActive'
            - $ref: '#/components/responses/InactiveMenuItem'
            - $ref: '#/components/responses/InsufficientStock'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/inventory/operation:
    get:
      tags: [inventory]
      summary: Получить список складских операций
      parameters:
        - in: query
          name: employee_id
          schema:
            type: string
        - in: query
          name: ingredient_id
          schema:
            type: string
        - in: query
          name: operation_type
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Список складских операций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOperationsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [inventory]
      summary: Создать ручную складскую операцию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperationRequest'
      responses:
        '200':
          description: Складская операция создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOperationResponse'
        '400':
          oneOf:
            - $ref: '#/components/responses/ValidationError'
            - $ref: '#/components/responses/ShiftNotActive'
            - $ref: '#/components/responses/InsufficientStock'
            - $ref: '#/components/responses/UnsupportedOperation'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/report/employees/excel:
    post:
      tags: [report]
      summary: Сформировать Excel-отчет по сотрудникам
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRangeRequest'
      responses:
        '200':
          $ref: '#/components/responses/ExcelFile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    IngredientID:
      in: path
      name: ingredient_id
      required: true
      schema:
        type: string
    MenuItemID:
      in: path
      name: menu_item_id
      required: true
      schema:
        type: string
...
  responses:
    ExcelFile:
      description: Excel файл
      content:
        application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
          schema:
            type: string
            format: binary
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: invalid request body
...
    ShiftNotActive:
      description: Нет активной смены
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: SHIFT_NOT_ACTIVE
              message: Активная смена отсутствует
    DuplicateName:
      description: Сущность с таким названием уже существует
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: DUPLICATE_NAME
              message: duplicate name
    IngredientUsedInRecipe:
      description: Ингредиент используется в рецептах
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INGREDIENT_USED_IN_RECIPE
              message: ingredient is used in recipes
    InactiveIngredient:
      description: Нельзя использовать неактивный ингредиент
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INACTIVE_INGREDIENT
              message: ingredient is inactive
    InactiveMenuItem:
      description: Нельзя использовать неактивный пункт меню
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INACTIVE_MENU_ITEM
              message: menu item is inactive
    InsufficientStock:
      description: Недостаточно остатка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INSUFFICIENT_STOCK
              message: insufficient stock
    UnsupportedOperation:
      description: Неподдерживаемый тип операции
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNSUPPORTED_OPERATION
              message: unsupported inventory operation
    NotFound:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: entity not found
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: internal server error
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    Pagination:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...
    Ingredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        unit:
          type: string
        current_stock:
          type: number
        is_active:
          type: boolean
        used_in_recipes_count:
          type: integer
    IngredientUsage:
      type: object
      properties:
        menu_item_id:
          type: string
        menu_item_name:
          type: string
        amount:
          type: number
    IngredientDetails:
      allOf:
        - $ref: '#/components/schemas/Ingredient'
        - type: object
          properties:
            used_in_recipes:
              type: array
              items:
                $ref: '#/components/schemas/IngredientUsage'
    ListIngredientsResponse:
      type: object
      properties:
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    GetIngredientResponse:
      type: object
      properties:
        ingredient:
          $ref: '#/components/schemas/IngredientDetails'
    CreateIngredientRequest:
      type: object
      required: [name, unit]
      properties:
        name:
          type: string
        unit:
          type: string
        initial_stock:
          type: number
          nullable: true
    CreateIngredientResponse:
      type: object
      properties:
        ingredient:
          $ref: '#/components/schemas/Ingredient'
    SetIngredientActivityRequest:
      type: object
      required: [is_active]
      properties:
        is_active:
          type: boolean
    SetIngredientActivityResponse:
      type: object
      properties:
        ingredient:
          $ref: '#/components/schemas/Ingredient'
    RecipeItem:
      type: object
      properties:
        ingredient_id:
          type: string
        ingredient_name:
          type: string
        unit:
          type: string
        amount:
          type: number
        ingredient_is_active:
          type: boolean
    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        is_active:
          type: boolean
        recipe:
          type: array
          items:
            $ref: '#/components/schemas/RecipeItem'
    ListMenuItemsResponse:
      type: object
      properties:
        menu_items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
    GetMenuItemResponse:
      type: object
      properties:
        menu_item:
          $ref: '#/components/schemas/MenuItem'
    CreateMenuItemRecipeItem:
      type: object
      required: [ingredient_id, amount]
      properties:
        ingredient_id:
          type: string
        amount:
          type: number
    CreateMenuItemRequest:
      type: object
      required: [name, price, recipe]
      properties:
        name:
          type: string
        price:
          type: number
        recipe:
          type: array
          items:
            $ref: '#/components/schemas/CreateMenuItemRecipeItem'
    CreateMenuItemResponse:
      type: object
      properties:
        menu_item:
          $ref: '#/components/schemas/MenuItem'
    UpdateMenuItemPriceRequest:
      type: object
      required: [price]
      properties:
        price:
          type: number
    UpdateMenuItemPriceResponse:
      type: object
      properties:
        menu_item:
          $ref: '#/components/schemas/MenuItem'
...
    OrderShift:
      type: object
      properties:
        id:
          type: string
        opened_at:
          type: string
          format: date-time
    OrderEmployee:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
    OrderItem:
      type: object
      properties:
        id:
          type: string
        menu_item_id:
          type: string
        menu_item_name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        total_price:
          type: number
    OrderInventoryOperation:
      type: object
      properties:
        id:
          type: string
        ingredient_id:
          type: string
        ingredient_name:
          type: string
        change_amount:
          type: number
        unit:
          type: string
        operation_type:
          type: string
        created_at:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        shift:
          $ref: '#/components/schemas/OrderShift'
        employee:
          $ref: '#/components/schemas/OrderEmployee'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        inventory_operations:
          type: array
          items:
            $ref: '#/components/schemas/OrderInventoryOperation'
        total_price:
          type: number
    ListOrdersResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    CreateOrderItemRequest:
      type: object
      required: [menu_item_id, quantity]
      properties:
        menu_item_id:
          type: string
        quantity:
          type: integer
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItemRequest'
    CreateOrderResponse:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
    InventoryOperationIngredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        unit:
          type: string
        current_stock:
          type: number
    InventoryOrderReference:
      type: object
      properties:
        id:
          type: string
    InventoryShiftReference:
      type: object
      properties:
        id:
          type: string
    InventoryEmployeeReference:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
    InventoryOperation:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        operation_type:
          type: string
        change_amount:
          type: number
        ingredient:
          $ref: '#/components/schemas/InventoryOperationIngredient'
        order:
          allOf:
            - $ref: '#/components/schemas/InventoryOrderReference'
          nullable: true
        shift:
          $ref: '#/components/schemas/InventoryShiftReference'
        employee:
          $ref: '#/components/schemas/InventoryEmployeeReference'
    ListOperationsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InventoryOperation'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateOperationRequest:
      type: object
      required: [ingredient_id, operation_type, amount]
      properties:
        ingredient_id:
          type: string
        operation_type:
          type: string
        amount:
          type: number
    CreateOperationResponse:
      type: object
      properties:
        operation:
          $ref: '#/components/schemas/InventoryOperation'
    TimeRangeRequest:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        employee_id:
          type: string
          nullable: true
...